+++ lib/main.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
- // import 'package:url_strategy/url_strategy.dart';
-import 'package:mws/app/routes/app_pages.dart'; // Keep this import
+import 'package:mws/app/routes/app_pages.dart';
 import 'package:mws/app/theme/app_theme.dart';
-import 'package:mws/utils/constants.dart';
+import 'package:mws/services/session_service.dart';
 
 void main() {
-  WidgetsFlutterBinding.ensureInitialized();
-  // setPathUrlStrategy();
+  // Initialize session service
+  Get.put(SessionService(), permanent: true);
+  
   runApp(const MyApp());
 }
 
@@ -17,11 +17,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
-      title: AppConstants.appTitle,
+      title: 'MWS DApp',
+      theme: AppTheme.darkTheme,
       initialRoute: AppPages.INITIAL,
       getPages: AppPages.routes,
-      theme: AppTheme.darkTheme,
       debugShowCheckedModeBanner: false,
     );
   }
 }
+